// src/app/models/propuesta.ts
export interface Propuesta {
    archivoEstudianteFecha: string;
  archivoEstudianteId: number;
  archivoEstudianteUrl: string;
  archivoFechaRevisor1: string | null;
  archivoFechaRevisor2: string | null;
  archivoIdRevisor1: number | null;
  archivoIdRevisor2: number | null;
  archivoUrlRevisor1: string | null;
  archivoUrlRevisor2: string | null;
  carrera: string;
  categoria: string;
  estadoAprobacion: string;
  estadoValidacion: string;
  id: number;
  nota1: number | null;
  nota2: number | null;
  observacion1: string | null;
  observacion2: string | null;
  observaciones: string | null;
  periodo: string;
  primerCarrera: string;
  primerEstuId: number;
  primerEstudiante: string;
  primerFacultad: string;
  primerRevisor: string;
  primerUsuaId: number;
  revisionId: number;
  revisionNumero: number;
  revisor1DocenId: number | null;
  revisor1UsuaId: number | null;
  revisor2DocenId: number | null;
  revisor2UsuaId: number | null;
  segundoCarrera: string | null;
  segundoEstuId: number | null;
  segundoEstudiante: string;
  segundoFacultad: string | null;
  segundoRevisor: string;
  segundoUsuaId: number | null;
  tareaId: string;
  tema: string;
  tercerCarrera: string | null;
  tercerEstuId: number | null;
  tercerEstudiante: string;
  tercerFacultad: string | null;
  tercerUsuaId: number | null;
  tipo: string;
  tutor: string;
  tutorDocenteId: number;
  tutorFacultad: string;
  tutorUsuaId: number;
}